+++ src/main/resources/db/migration/V3__Create_question_table.sql
@@ -1,9 +1,9 @@
 create table question
 (
-	id int auto_increment,
+	id bigint auto_increment,
 	title varchar(50),
 	description text,
-	creator int,
+	creator bigint,
 	comment_count int default 0,
 	view_count int default 0,
 	like_count int default 0,
